Moved the Animation functions into the components folder where they are a better fit.
