+++ v3/src/checksum.js
@@ -1,4 +1,4 @@
 var CHECKSUM = {
-build: '253a8370-2f8f-11e7-8cb0-75ca1491836b'
+build: '77d3f640-2f92-11e7-94cb-df78655b45e3'
 };
 module.exports = CHECKSUM;
