+++ v3/src/components/animation/Pause.js
