+++ v3/src/components/animation/Paused.js
