+++ v3/src/components/animation/Progress.js
