+++ v3/src/components/animation/Remove.js
@@ -4,8 +4,6 @@ var Remove = function (event)
 
     if (this.isPlaying && event.key === this.currentAnim.key)
     {
-        console.log('animationRemoved', event.key);
-
         this.stop();
 
         var sprite = this.parent;
