+++ v3/src/components/animation/Stop.js
