+++ v3/src/components/animation/TimeScale.js
