+++ v3/src/components/animation/TotalFrames.js
