+++ v3/src/components/animation/TotalProgress.js
