+++ v3/src/components/animation/Yoyo.js
